requirement.txt
numpy
pandas
GitPython
hydra-core
tqdm
tensorboard
scipy
Cython==0.29.35
torch==2.3.0
torch_geometric==2.4.0
torch-scatter==2.1.2